+++ src/Singleton.php
@@ -103,29 +103,16 @@ if ( ! class_exists( 'Fragen\\Singleton' ) ) {
 		}
 
 		/**
-		 * Returns error message for not finding a class.
+		 * Logs and returns error message for not finding a class.
 		 *
 		 * @param string $class_name
 		 */
 		private static function get_error( $class_name ) {
-			$error     = '<tr><td><pre><strong>PHP Fatal:</strong> Undefined class "' . $class_name . '"</pre></td></tr>';
-			$Exception = new \Exception( $error );
-			$trace     = $Exception->getTrace();
-			$trace     = array_reverse( $trace );
-			$message   = $Exception->getMessage();
-			$message  .= '<tr><td><pre>PHP Stack Trace:</pre></td></tr>';
-			$i         = 0;
-			foreach ( $trace as $err ) {
-				$i++;
-				$message .= '<tr><td><pre>';
-				$message .= sprintf(
-					$i . '. %1$s called from %2$s:%3$s',
-					'<strong>' . $err['class'] . $err['type'] . $err['function'] . '()</strong>',
-					'<strong>' . $err['file'] . '</strong>',
-					$err['line'] . '</pre></td></tr>'
-				);
-			}
-			die( '<table>' . $message . '</table>' );
+			$message   = "PHP Fatal error: Undefined class - '{$class_name}'\nPHP Stack trace:\n";
+			$Exception = new \Exception();
+			$trace     = $Exception->getTraceAsString();
+			error_log( $message . $trace );
+			die( "<pre><strong>{$message}</strong>{$trace}</pre>" );
 		}
 	}
 }